Furniture recolor: i have made an another scene wich appears only if the transform controls has an object attached, and the user clicks the painter-button.
in the scene you can click on parts of the furniture and the colorpicker picks the color of the part, you can modify that. If u done modifiing your furniture, you can click ok to apply the modifications.

I needed a copy constructor like structure for the furniture, the .clone doesn't work because the model.userdata.root was set to the base furniture, so the jsonify stucked in a circle...
solution: i had made a large modification for the furniture constructor wich clones every attribute of the other furniture (if its not null) and loads a new model in based on the catalogItem, than applies colors of the original.
so its basically clones everithing except the model, wich is manually loaded in again.